Subject: Added responsive behavior to NavMenu
In this commit, we have made significant updates to the `MainLayout.razor` and `NavMenu.razor` files to improve the responsiveness of the navigation menu.

In `MainLayout.razor`, a new boolean variable `isSmallScreen` has been introduced to check for small screen sizes. We have added a new method `OnMediaQueryChange` to update this variable based on the media query. Another method `HandleMenuItemClick` has been added to collapse the sidebar when a menu item is clicked on a small screen. The `NavMenu` component now has a new parameter `OnMenuItemClicked` which is set to the `HandleMenuItemClick` method.

In `NavMenu.razor`, we have included a new `OnMenuItemClicked` parameter. The `RadzenPanelMenuItem` component now has a `Click` event which triggers the `OnMenuItemClick` method. This method checks if the `OnMenuItemClicked` delegate has been assigned and if so, it invokes the delegate. This allows the parent component to handle the menu item click event. The rendering of `RadzenPanelMenuItem` has been updated to only render children menu items if they match the search text.